FastApi-server/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
google-generativeai
apscheduler
python-dotenv
